Fixed some type conversion issues.  The Blitter scaling of the main menu leaves a small gap between the scaled and scrolling images in OpenGL mode, though.

+++ src/display/drawing_context.cxx
@@ -378,13 +378,13 @@ DrawingContext::get_clip_rect()
 float
 DrawingContext::get_width() const
 {
-  return CL_Display::get_width();
+  return (float)CL_Display::get_width();
 }
 
 float
 DrawingContext::get_height() const
 {
-  return CL_Display::get_height();  
+  return (float)CL_Display::get_height();  
 }
 
 void
